Implementing the popup confirmation for deleting a favorite and implementing the gesture

+++ www/js/controllers.js
@@ -216,21 +216,29 @@ angular.module('conFusion.controllers', [])
     $scope.baseURL = baseURL;
     $scope.leadership = corporateFactory.query();
   }])
-  .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
-
+  .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
     $scope.baseURL = baseURL;
     $scope.shouldShowDelete = false;
+
+    $ionicLoading.show({
+      template: '<ion-spinner></ion-spinner> Loading...'
+    });
 
     $scope.favorites = favoriteFactory.getFavorites();
 
     $scope.dishes = menuFactory.getDishes().query(
       function (response) {
         $scope.dishes = response;
+        $timeout(function () {
+          $ionicLoading.hide();
+        }, 1000);
       },
       function (response) {
         $scope.message = "Error: " + response.status + " " + response.statusText;
+        $timeout(function () {
+          $ionicLoading.hide();
+        }, 1000);
       });
-    console.log($scope.dishes, $scope.favorites);
 
     $scope.toggleDelete = function () {
       $scope.shouldShowDelete = !$scope.shouldShowDelete;
@@ -238,8 +246,20 @@ angular.module('conFusion.controllers', [])
     };
 
     $scope.deleteFavorite = function (index) {
+      var confirmPopup = $ionicPopup.confirm({
+        title: 'Confirm Delete',
+        template: 'Are you sure you want to delete this item?'
+      });
 
-      favoriteFactory.deleteFromFavorites(index);
+      confirmPopup.then(function (res) {
+        if (res) {
+          console.log('Ok to delete');
+          favoriteFactory.deleteFromFavorites(index);
+        } else {
+          console.log('Canceled delete');
+        }
+      });
+
       $scope.shouldShowDelete = false;
 
     }}])
